Fix issue for long numbers.

+++ java/src/com/dev/factorial/Factorial.java
@@ -2,7 +2,7 @@ package com.dev.factorial;
 
 
 public class Factorial {
-    public static int calculate(int n) {
+    public static long calculate(long n) {
         if (n<=1) return 1;
         return calculate(n-1) * n;
     }
